+++ mgm-hurry
@@ -1,6 +1,15 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 # -* vim: syntax=python -*-
+#                                       _
+#                                      | |
+#  _ __ ___    __ _  _ __ ___   ______ | |__   _   _  _ __  _ __  _   _
+# | '_ ` _ \  / _` || '_ ` _ \ |______|| '_ \ | | | || '__|| '__|| | | |
+# | | | | | || (_| || | | | | |        | | | || |_| || |   | |   | |_| |
+# |_| |_| |_| \__, ||_| |_| |_|        |_| |_| \__,_||_|   |_|    \__, |
+#              __/ |                                               __/ |
+#             |___/                                               |___/
+#
 # --- ↑↓ Do not remove these libs ↑↓ -----------------------------------------------------------------------------------
 import glob
 import json
@@ -17,8 +26,8 @@ from InquirerPy.validator import NumberValidator
 from art import tprint
 
 from user_data.mgm_tools.mgm_hurry.FreqtradeCli import FreqtradeCli
-from user_data.mgm_tools.mgm_hurry.LeetLogger import get_logger
 from user_data.mgm_tools.mgm_hurry.MoniGoManiCli import MoniGoManiCli
+from user_data.mgm_tools.mgm_hurry.MoniGoManiLogger import get_logger
 # ---- ↑ Do not remove these libs ↑ ------------------------------------------------------------------------------------
