+++ tests/test_mgm_hurry_freqtrade_cli.py
@@ -134,7 +134,7 @@ def __get_instance(dir, use_logger=True):
     logger = None
 
     if use_logger is True:
-        from user_data.mgm_tools.mgm_hurry.LeetLogger import get_logger
+        from user_data.mgm_tools.mgm_hurry.MoniGoManiLogger import get_logger
         logger = get_logger()
 
     fc = FreqtradeCli(dir, logger)
