+++ user_data/mgm_tools/mgm_hurry/FreqtradeCli.py
@@ -1,5 +1,15 @@
 # -*- coding: utf-8 -*-
 # -* vim: syntax=python -*-
+#
+# ______                     _                     _        _____  _  _
+# |  ___|                   | |                   | |      /  __ \| |(_)
+# | |_    _ __   ___   __ _ | |_  _ __   __ _   __| |  ___ | /  \/| | _
+# |  _|  | '__| / _ \ / _` || __|| '__| / _` | / _` | / _ \| |    | || |
+# | |    | |   |  __/| (_| || |_ | |   | (_| || (_| ||  __/| \__/\| || |
+# \_|    |_|    \___| \__, | \__||_|    \__,_| \__,_| \___| \____/|_||_|
+#                        | |
+#                        |_|
+#
 # --- ↑↓ Do not remove these libs ↑↓ -----------------------------------------------------------------------------------
 import os
 import tempfile
