--- user_data/mgm_tools/mgm_hurry/LeetLogger.py
@@ -1,17 +1,0 @@
-# -*- coding: utf-8 -*-
-# -* vim: syntax=python -*-
-# --- ↑↓ Do not remove these libs ↑↓ -----------------------------------------------------------------------------------
-import logging
-import coloredlogs
-# ---- ↑ Do not remove these libs ↑ ------------------------------------------------------------------------------------
-
-
-def get_logger():
-    """
-    Let's Log and Roll
-    """
-    logging.basicConfig(format='= MGM-HURRY = %(levelname)s: %(message)s', level=logging.DEBUG)
-    logger = logging.getLogger(__name__)
-    coloredlogs.install(level=logging.DEBUG, logger=logger)
-
-    return logger
