+++ user_data/mgm_tools/mgm_hurry/MoniGoManiLogger.py
@@ -1,0 +1,27 @@
+# -*- coding: utf-8 -*-
+# -* vim: syntax=python -*-
+#
+# ___  ___               _  _____        ___  ___               _  _
+# |  \/  |              (_)|  __ \       |  \/  |              (_)| |
+# | .  . |  ___   _ __   _ | |  \/  ___  | .  . |  __ _  _ __   _ | |      ___    __ _   __ _   ___  _ __
+# | |\/| | / _ \ | '_ \ | || | __  / _ \ | |\/| | / _` || '_ \ | || |     / _ \  / _` | / _` | / _ \| '__|
+# | |  | || (_) || | | || || |_\ \| (_) || |  | || (_| || | | || || |____| (_) || (_| || (_| ||  __/| |
+# \_|  |_/ \___/ |_| |_||_| \____/ \___/ \_|  |_/ \__,_||_| |_||_|\_____/ \___/  \__, | \__, | \___||_|
+#                                                                                 __/ |  __/ |
+#                                                                                |___/  |___/
+#
+# --- ↑↓ Do not remove these libs ↑↓ -----------------------------------------------------------------------------------
+import logging
+import coloredlogs
+# ---- ↑ Do not remove these libs ↑ ------------------------------------------------------------------------------------
+
+
+def get_logger():
+    """
+    Let's Log and Roll
+    """
+    logging.basicConfig(format='= MGM-HURRY = %(levelname)s: %(message)s', level=logging.DEBUG)
+    logger = logging.getLogger(__name__)
+    coloredlogs.install(level=logging.DEBUG, logger=logger)
+
+    return logger
